requirements.txt
absl-py==1.4.0
amqp==5.1.1
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.0
celery==5.2.7
certifi==2020.6.20
cffi==1.15.1
channels==3.0.3
channels-redis==4.0.0
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmdtest==0.32+git
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
confluent-kafka==2.0.2
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.1
cupshelpers==1.0
cycler==0.11.0
daphne==3.0.2
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==3.2
django-3-jet==1.0.8
django-4-jet==1.0.9
django-allauth==0.53.1
django-celery-beat==2.4.0
django-channels==0.7.0
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-material-admin==1.8.6
django-paypal==2.0
django-rest-auth==0.9.5
django-timezone-field==5.0
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.3
filelock==3.9.0
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.39.3
fqdn==1.5.1
future==0.18.2
gast==0.4.0
gdown==4.7.1
geojson==3.0.1
gmaps==0.9.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
googlemaps==4.10.0
grpcio==1.54.0
gyp==0.1
h5py==3.8.0
hdfs==2.7.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==22.10.0
ipykernel==6.22.0
ipython==8.13.1
ipython_genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.1
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterthemes==0.20.0
kafka-python==2.0.2
keras==2.12.0
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lesscpy==0.15.1
libclang==16.0.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.1.0
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.5
nbclassic==0.5.6
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
passlib==1.7.4
paypal==1.2.5
paypalrestsdk==1.13.1
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.1
Pillow==9.0.1
platformdirs==2.6.2
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.22.4
psutil==5.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyDrive==1.3.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-crontab==2.7.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.7
python-dotenv==0.19.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.2
redis==4.5.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.1
Send2Trash==1.8.2
service-identity==21.1.0
six==1.16.0
sklearn==0.0.post4
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.3
ssh-import-id==5.11
stack-data==0.6.2
stevedore==4.1.1
systemd-python==234
ta==0.10.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
texttable==1.6.4
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
ttystatus==0.38
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.0
